#10 add sidebar test
