+++ components/Navigation/index.ts
@@ -1,0 +1,4 @@
+import NavBar from "./NavBar";
+import SideBar from "./SideBar";
+
+export { NavBar, SideBar };
