--- components/index.ts
@@ -1,7 +1,0 @@
-import NavBar from "./Navigation/NavBar";
-import SideBar from "./Navigation/SideBar";
-
-export {
-    NavBar,
-    SideBar
-}
