+++ jest.setup.js
@@ -1,1 +1,1 @@
-import "@testing-library/jest-dom";
+import "@testing-library/jest-dom/extend-expect";
